requirements.txt
aadict==0.2.3
appdirs==1.4.4
asset==0.6.13
async-timeout==4.0.2
bidict==0.22.0
bleach==5.0.0
blinker==1.4
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.12
ciso8601==2.2.0
click==8.1.3
cryptography==37.0.2
decorator==4.4.2
deepl==1.8.0
defusedxml==0.7.1
dnspython==2.2.1
eventlet==0.33.1
ffmpeg-python==0.2.0
Flask==2.1.2
Flask-Assets==2.0
Flask-Caching==2.0.2
Flask-Mail==0.9.1
Flask-SocketIO==5.2.0
Flask-WTF==1.0.1
fonttools==4.38.0
fpdf2==2.5.7
fs==2.4.16
furl==2.1.3
future==0.18.3
globre==0.1.5
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.21.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
less==0.0.1
lxml==4.9.1
MarkupSafe==2.1.1
moviepy==1.0.3
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.2
orderedmultidict==1.0.1
packaging==21.3
pandas==1.4.2
pikepdf==5.1.5
Pillow==9.3.0
pip-autoremove==0.10.0
proglog==0.1.10
protobuf==4.21.11
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyjsonq==1.0.2
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.2
python-slugify==6.1.2
python-socketio==5.6.0
pytz==2022.1
qrcode==7.3.1
rcssmin==1.1.0
redis==4.5.4
requests==2.28.0
requests-oauthlib==1.3.1
rjsmin==1.2.0
rsa==4.9
six==1.16.0
svg.path==6.2
text-unidecode==1.3
tqdm==4.64.0
triad==0.6.3
uritemplate==4.1.1
urllib3==1.26.9
webassets==2.0
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.8.0
